moved from_root & to_root to HistogramByPileUpCollection
